+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-02-13
+
+### Added
+
+- Pipeline for matching datasets using the CompanyMatching service
 
 ## 2020-02-11